Ok, I think I know where I can go to make this better, and I'll work on it a little after I get 3 hrs. Basically, I will take the training data, iterate through it all, and give each String a corresponding binId in another 2d list called binMap. This will allow the algorithm to clone into sublists to pass on into further calls, and we can get rid of rows called examples (part of the process, txtbook pg 703 the line saying exs <-{e:eEexamples and e.A = vk}) then when we want to know what String is in a data slot, we go to that attribute's bin list, and look up that binId is, which we use to label tree branches.
The key that seems to be appearing in figuring out this alg is to keep multiple 2d and 3d lists updated, as the index of one will allow you to look up a corresponding value in another. NOTE: since the classification attribute has been seperated, the examples and parent_examples parameters need an extra parameter that will hold examplesClasses and parentClasses. My brain is dead, I'll be up at 8:30/9